updating repo and attempting to add autodiff for thermo module
