+++ oscps-lib/src/stream.rs
@@ -1,7 +1,7 @@
 //! # Connector
 //!
 
-use crate::thermodynamics::StreamThermoState; 
+use crate::thermodynamics::{ThermodynamicConstants, ThermoPackage}; 
 
 /// # Stream
 /// 
@@ -35,3 +35,85 @@ impl Stream {
         }
     }
 }
+
+#[allow(dead_code)]
+/// Species list
+pub struct ComponentData {
+    /// Chemical species
+    pub chemical_species: Chemical, // will contain intrinsic properties of species
+    /// Mass quantity
+    pub mass_quantity: Mass,
+    /// Molar quantity
+    pub molar_quantity: AmountOfSubstance,
+    ///volumetric quantity
+    pub vol_quantity: Volume,
+    /// partial pressure
+    pub partial_pressure : Pressure,
+}
+
+#[allow(dead_code)]
+/// # StreamThermoState
+/// 
+/// This struct will be used for performing thermodynamic calculations for the streams in the flow
+/// diagram.
+pub struct StreamThermoState {
+    /// Pressure of the state.
+    pub pressure: Option<Pressure>,                    // pressure
+    /// Temperature of the state.
+    pub temperature: Option<ThermodynamicTemperature>, // temperature
+    /// List of mole fractions.
+    pub mass_list: Vec<ComponentData>,//Information about each component within stream
+    /// Total Mass
+    pub total_mass : Option<Mass>, // total mass in stream
+    /// Total Moles
+    pub total_mol : Option<AmountOfSubstance>, // total moles in stream
+    /// Total Volume
+    pub total_volume : Option<Volume>, // total volume in stream
+    ///Thermo Package
+    pub thermodynamic_package : Option<Box<dyn ThermoPackage>> // thermodynamics package 
+}
+
+
+#[allow(dead_code)]
+/// Implementation of StreamThermoState
+/// This struct holds the functionality to perform thermodynamic calculations for streams
+impl StreamThermoState {
+    /// Constructor for creating a StreamThermoState
+    pub fn new() -> Self {
+        StreamThermoState {
+            pressure : None,
+            temperature : None,
+            mass_list : vec![],
+            total_mass : None,
+            total_mol : None,
+            total_volume : None,
+            thermodynamic_package : None
+        }
+    }
+    /// Public function to execute the calculations for determining the thermodynamic state for the
+    /// stream. Dependence on the Thermodynamic Packages.
+    pub fn execute_thermo_calcs(&mut self) -> Self {
+
+    }
+    /// this function will return the total mass for an individual stream
+    fn calc_total_mass(&mut self) -> Mass {
+        let mut mass_sum  = 0.0;
+        for chem in &self.mass_list {
+            mass_sum += chem.mass_quantity.get::<mass::kilogram>();
+        }
+        self.total_mass = Some(Mass::new::<mass::kilogram>(mass_sum));
+        
+        self.total_mass.unwrap()
+    }
+    /// this function will return the total moles for an individual stream
+    fn calc_total_moles(&mut self) -> AmountOfSubstance {
+        let mut mole_sum  = 0.0;
+        for chem in &self.mass_list {
+            mole_sum += chem.molar_quantity.get::<amount_of_substance::mole>();
+        }
+        self.total_mol = Some(AmountOfSubstance::new::<amount_of_substance::mole>(mole_sum));
+        
+        self.total_mol.unwrap()
+    }
+}
+
